Fixes a bug in gumtree that causes non-deterministic results.
As the similarity between two entries is not unique it is possible
that across different runs of gumtree on the same changed file the order
in ambiguousList changes. Due to srcIgnored and dstIgnored in
filterMappings this can cause a change in the mapping. The added ID
criteria enforces an absolute order. Of course other absolute orders
would also be valid.